make folders kebab-case, change schema table names
