+++ src/indexer/actions/index.ts
@@ -7,13 +7,13 @@ export * from './staking/bonded';
 export * from './staking/unbonded';
 export * from './staking/withdrawn';
 export * from './staking/payee';
-export * from './parachainStaking/reward';
-export * from './parachainStaking/compound';
-export * from './parachainStaking/delegation';
-export * from './parachainStaking/delegationRevoked';
-export * from './parachainStaking/delegationIncreased';
-export * from './parachainStaking/delegationDecreased';
-export * from './parachainStaking/delegationKicked';
-export * from './parachainStaking/candidateLeft';
-export * from './parachainStaking/candidateBondedMore';
-export * from './parachainStaking/candidateBondedLess';
+export * from './parachain-staking/reward';
+export * from './parachain-staking/compound';
+export * from './parachain-staking/delegation';
+export * from './parachain-staking/delegationRevoked';
+export * from './parachain-staking/delegationIncreased';
+export * from './parachain-staking/delegationDecreased';
+export * from './parachain-staking/delegationKicked';
+export * from './parachain-staking/candidateLeft';
+export * from './parachain-staking/candidateBondedMore';
+export * from './parachain-staking/candidateBondedLess';
