+++ src/indexer/actions/parachain-staking/candidateBondedMore.ts
@@ -1,7 +1,7 @@
-import { Account, ParachainStakingCandidatesUnbond, Staker } from '@/model';
+import { Account, ParachainStakingCandidatesBonded, Staker } from '@/model';
 import { Action, ActionContext } from '../base';
 
-interface ParachainCandidateBondedLessData {
+interface ParachainCandidateBondedMoreData {
   id: string;
   amount: bigint;
   account: () => Promise<Account>;
@@ -9,12 +9,12 @@ interface ParachainCandidateBondedLessData {
   extrinsicHash: string;
 }
 
-export class ParachainCandidateBondedLessAction extends Action<ParachainCandidateBondedLessData> {
+export class ParachainCandidateBondedMoreAction extends Action<ParachainCandidateBondedMoreData> {
   protected async _perform(ctx: ActionContext): Promise<void> {
     const account = await this.data.account();
     const staker = await this.data.staker();
 
-    const candidateUnbonded = new ParachainStakingCandidatesUnbond({
+    const candidateBondedMore = new ParachainStakingCandidatesBonded({
       id: this.data.id,
       blockNumber: this.block.height,
       timestamp: new Date(this.block.timestamp ?? 0),
@@ -24,9 +24,9 @@ export class ParachainCandidateBondedLessAction extends Action<ParachainCandidat
       amount: this.data.amount,
     });
 
-    staker.totalBonded -= this.data.amount;
+    staker.totalBonded += this.data.amount;
 
-    await ctx.store.insert(candidateUnbonded);
+    await ctx.store.insert(candidateBondedMore);
     await ctx.store.upsert(staker);
   }
 }
