+++ src/indexer/actions/parachain-staking/candidateLeft.ts
@@ -1,4 +1,4 @@
-import { Account, ParachainStakingCandidatesWithdraw, Staker } from '@/model';
+import { Account, ParachainStakingCandidatesWithdrawed, Staker } from '@/model';
 import { Action, ActionContext } from '../base';
 
 interface ParachainCandidateLeftData {
@@ -14,7 +14,7 @@ export class ParachainCandidateLeftAction extends Action<ParachainCandidateLeftD
     const account = await this.data.account();
     const staker = await this.data.staker();
 
-    const candidateUnbondAndWithdraw = new ParachainStakingCandidatesWithdraw({
+    const candidateUnbondAndWithdraw = new ParachainStakingCandidatesWithdrawed({
       id: this.data.id,
       blockNumber: this.block.height,
       timestamp: new Date(this.block.timestamp ?? 0),
