+++ src/indexer/pallets/parachain-staking/events/candidateBondedLess.ts
