+++ src/indexer/pallets/parachain-staking/events/candidateBondedMore.ts
