+++ src/indexer/pallets/parachain-staking/events/candidateLeft.ts
