+++ src/indexer/pallets/parachain-staking/events/compounded.ts
@@ -1,18 +1,17 @@
-import { EnsureAccount, EnsureStaker, ParachainDelegationAction } from '@/indexer/actions';
+import { EnsureAccount, EnsureStaker, ParachainCompoundAction } from '@/indexer/actions';
 import { EventPalletHandler, IEventHandlerParams, IHandlerOptions } from '@/indexer/pallets/handler';
-import { buildParachainStakingExtrinsicHash } from '@/indexer/pallets/parachainStaking/utils';
+import { buildParachainStakingExtrinsicHash } from '@/indexer/pallets/parachain-staking/utils';
 import { IEventPalletDecoder } from '@/indexer/types';
 import { Account, Staker } from '@/model';
 
-export interface IParachainDelegationEventDecoder
-  extends IEventPalletDecoder<{ delegator: string; amount: bigint; stash: string; delegatorPosition: { __kind: string }; autoCompundPercent?: number }> {}
+export interface IParachainCompoundEventDecoder extends IEventPalletDecoder<{ stash: string; delegator: string; amount: bigint }> {}
 
-interface IDelegationEventPalletSetup {
-  decoder: IParachainDelegationEventDecoder;
+interface ICompoundEventPalletSetup {
+  decoder: IParachainCompoundEventDecoder;
 }
 
-export class ParachainDelegationEventPalletHandler extends EventPalletHandler<IDelegationEventPalletSetup> {
-  constructor(setup: IDelegationEventPalletSetup, options: IHandlerOptions) {
+export class ParachainCompoundEventPalletHandler extends EventPalletHandler<ICompoundEventPalletSetup> {
+  constructor(setup: ICompoundEventPalletSetup, options: IHandlerOptions) {
     super(setup, options);
   }
 
@@ -26,7 +25,7 @@ export class ParachainDelegationEventPalletHandler extends EventPalletHandler<ID
 
     const accountDef = ctx.store.defer(Account, stakerId);
     const delegatorDef = ctx.store.defer(Account, delegatorId);
-    const stakerDef = ctx.store.defer(Staker, stakerId);
+    const stakerDef = ctx.store.defer(Staker, { id: stakerId });
 
     queue.push(
       new EnsureAccount(block.header, event.extrinsic, {
@@ -44,15 +43,13 @@ export class ParachainDelegationEventPalletHandler extends EventPalletHandler<ID
         staker: () => stakerDef.get(),
         account: () => accountDef.getOrFail(),
       }),
-      new ParachainDelegationAction(block.header, event.extrinsic, {
+      new ParachainCompoundAction(block.header, event.extrinsic, {
         id: event.id,
         account: () => accountDef.getOrFail(),
         staker: () => stakerDef.getOrFail(),
         delegator: () => delegatorDef.getOrFail(),
         amount: data.amount,
         extrinsicHash: buildParachainStakingExtrinsicHash(block.header.height, event.index, event.extrinsicIndex),
-        delegatorPosition: data.delegatorPosition.__kind,
-        autoCompundPercent: data.autoCompundPercent,
       })
     );
   }
