+++ src/indexer/pallets/parachain-staking/events/delegationRevoked.ts
